Support > 2GiB seeks in adb_lseek.
Bug: http://b/112003354
Test: builds

+++ adb/client/commandline.cpp
@@ -887,7 +887,7 @@ static int adb_sideload_host(const char* filename) {
             to_write = sb.st_size - offset;
         }
 
-        if (adb_lseek(package_fd, offset, SEEK_SET) != static_cast<int>(offset)) {
+        if (adb_lseek(package_fd, offset, SEEK_SET) != static_cast<int64_t>(offset)) {
             fprintf(stderr, "adb: failed to seek to package block: %s\n", strerror(errno));
             return -1;
         }
